DiabloUI: Move art and text code out of diablo.cpp
