+++ SourceX/.clang-format
@@ -8,4 +8,5 @@
 	UseTab: ForIndentation,
 	SortIncludes: false,
 	NamespaceIndentation: None,
+	FixNamespaceComments: true,
 }
