+++ SourceX/DiabloUI/art.h
@@ -27,6 +27,7 @@ struct Art {
 	}
 };
 
-void LoadArt(char *pszFile, Art *art, int frames = 1, PALETTEENTRY *pPalette = NULL);
+void LoadArt(const char *pszFile, Art *art, int frames = 1, PALETTEENTRY *pPalette = NULL);
+void LoadMaskedArt(const char *pszFile, Art *art, int frames = 1, int mask = 250);
 
-}
+} // namespace dvl
