+++ SourceX/DiabloUI/text_draw.h
@@ -1,0 +1,9 @@
+#pragma once
+
+#include "devilution.h"
+
+namespace dvl {
+
+void DrawArtStr(const char *text, const SDL_Rect &rect, int flags, bool drawTextCursor = false);
+
+} // namespace dvl
